% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree_function_extensions.R
\name{add_clades_to_tree}
\alias{add_clades_to_tree}
\title{Highlight clades over a ggtree object given a named vector of internal nodes and labels to draw}
\usage{
add_clades_to_tree(
  drawn_tree,
  clade_labels,
  highlight_clades = TRUE,
  aligned = TRUE,
  extend_to_value = 0,
  highlight_alpha = 0.4,
  label_offset = 0,
  gradient = F,
  gradient_direction = "tr"
)
}
\arguments{
\item{drawn_tree}{ggtree object for labels to be added to}

\item{clade_labels}{Named vector of internal node and label (as names), e.g. output from micro.gen.extra::get_clade_nodes()}

\item{highlight_clades}{TRUE/FALSE as to whether to include the ggtree::geom_hilight [Default = TRUE]}

\item{aligned}{TRUE/FALSE as to whether the ggtree::geom_hilights should align [Default = TRUE]}

\item{extend_to_value}{Value to extend the ggtree::geom_hilight to. Float. [Default = 0.0]}

\item{highlight_alpha}{Value to set the alpha of the ggtree::geom_hilight. Set between 0 and 1. [Default = 0.4]}

\item{label_offset}{Value to offset the labels of the clade_labels. [Default = F]}

\item{gradient}{T/F for whether hilighted box should have a gradient (one of "rt" or "tr" [root-to-tip or tip-to-root]) [Default = "tr"]}

\item{gradient_drcton}{Gradient direction [Default = 0.0]}
}
\value{
ggtree object with clades (corresponding to the internal nodes supplied) have been highlighted and labelled.
}
\description{
Given a named vector of internal tree nodes and labels (e.g. cluster number / name), move through each cluster name
and apply geom_hilight and geom_cladelabel from ggtree onto the tree, with alternating shades of grey from top
to bottom.
}
\examples{
add_clades_to_tree(ggtree_object, clade_labels_vector)
}
