% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree_function_extensions.R
\name{collapse_tree}
\alias{collapse_tree}
\title{Collapse a set of nodes in a ggtree object}
\usage{
collapse_tree(tree, nodes, type = "max", colours = "none")
}
\arguments{
\item{tree}{Tree object}

\item{nodes}{List of internal nodes to collapse}

\item{type}{Type of collapsing to run - "min", "max" or "mixed". [Default = "max"]}

\item{colours}{List of hex colours to colour the collapsed cones, order of the colours corresponds
with the list of nodes given [Default = "none"]}
}
\value{
Tree object with requested tips pruned out
}
\description{
Given a list of nodes, collapse these nodes in a ggtree object
}
\examples{
remove_tips_from_tree(tree_object, list_of_tips)
}
