% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoplotR_extensions.R
\name{create_blast_comparisons}
\alias{create_blast_comparisons}
\title{Create blast comparisons for synteny plots}
\usage{
create_blast_comparisons(
  blast_order,
  fasta_dir,
  blast_results_dir = "blast_results",
  overwrite = FALSE,
  color_scheme_name = "red_blue"
)
}
\arguments{
\item{blast_order}{List of the fasta files to be compared, not including the ".fasta" extension.}

\item{fasta_dir}{Directory where the fasta files are kept. If in the working directory, set this to "."}

\item{blast_results_dir}{Directory to put the blast results into [Default = "blast_results"]}

\item{overwrite}{TRUE/FALSE to overwrite a previously existing blast_results_dir [Default = FALSE]}

\item{color_scheme_name}{color scheme to be used for the genoplotR comparison [Default = "red_blue"]}
}
\value{
a dna_comparisons list for genoplotR in addition to a directory with blast comparisons for a list of fasta sequences.
Comparisons will be sequential. e.g. sequence_1 vs sequence_2, sequence_2 vs sequence_3, sequence_3 vs sequence_4
}
\description{
Create blast comparisons for synteny plots provided a list of fasta sequences,
These can then be used as a readable format for genoplotR for synteny plots.
The order of the fasta sequences will be the order of the genomic regions to be compared in the synteny plot
from top to bottom.
blastn and makeblastdb
}
\examples{
create_blast_comparisons(blast_order_list, "dir_with_fasta_files")
}
