% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pan_genome_extensions.R
\name{create_upset_dfs}
\alias{create_upset_dfs}
\title{Create UpSetR dataframes from a gene_presence_absence.Rtab file}
\usage{
create_upset_dfs(
  gene_presence_absence_file,
  groups,
  levels = c(0, 0.15, 0.95, 1),
  labels = c("Cloud", "Shell", "Core")
)
}
\arguments{
\item{gene_presence_absence_file}{gene_presence_absence.Rtab file from Roary or Panaroo}

\item{groups}{Dataframe of the groups with two columns, one named "strain" listing strain names, and the other named "group" listing the group/lineage etc that each strain belongs to}

\item{levels}{Levels to define the pan-genome [Default = c(0,0.15,0.95,1)]}

\item{labels}{Labels for groups created between each of the values in levels. Should be of length : (length(levels) - 1).
[Default = c("Cloud","Shell","Core")]}
}
\value{
UpsetR dataframe ready for plotting
}
\description{
Create UpSetR dataframes from a gene_presence_absence.Rtab file, with the different groups to create the
sets being defined by clusters - e.g. phylogroups/lineages calculated by fastANI/PopPUNK/fastbaps etc
}
\examples{
create_upset_dfs(gene_presence_absence.Rtab, list_of_groups)
}
