% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_parsing.R
\name{fasta_from_single_gff}
\alias{fasta_from_single_gff}
\title{Get fasta sequence from single-contig gff3 file}
\usage{
fasta_from_single_gff(gff_file, fasta_dir, clean_up = TRUE)
}
\arguments{
\item{gff_file}{Gff file}

\item{fasta_dir}{Directory to put the fasta file into [Default = "."]}

\item{clean_up}{TRUE/FALSE - clean up intermediate files. [Default = TRUE]}
}
\value{
fasta sequence as string
}
\description{
Get fasta sequence from a single-contig gff3 file and return it as a string
}
\examples{
fasta_from_single_gff("input.gff")
}
