% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_parsing.R
\name{fasta_lengths_from_gff_list}
\alias{fasta_lengths_from_gff_list}
\title{Get fasta sequence from several single-contig gff3 file}
\usage{
fasta_lengths_from_gff_list(
  gff_list,
  gff_names,
  fasta_dir = "temp_fasta_files",
  clean_up = TRUE
)
}
\arguments{
\item{gff_list}{a list of the gff_files to be used - with full / relative file path}

\item{gff_names}{a list of names to be used for each file in gff_list, in the same order}

\item{fasta_dir}{Directory to put the fasta files in to [Default = "."]}

\item{clean_up}{TRUE/FALSE - clean up intermediate files. [Default = TRUE]}
}
\value{
List of fasta sequences
}
\description{
Get fasta sequence from several single-contig gff3 file and return it as a list of strings
}
\examples{
fasta_lengths_from_gff_list(List of gffs, names of the gff files)
}
