% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_parsing.R
\name{fasta_to_df}
\alias{fasta_to_df}
\title{Read in sequence data and fasta header from a fasta file into a tidy dataframe}
\usage{
fasta_to_df(fasta_file, type = "AA")
}
\arguments{
\item{fasta_file}{Input fasta file}

\item{type}{One of "DNA" or "AA" - default = "AA"}
}
\value{
Data frame with sequences and headers from a fasta file (plus molecular weight and pI if fasta file is protein sequences)
}
\description{
Read in sequence data and fasta header from a fasta file into a tidy dataframe.
Specify whether amino acid or DNA sequence in fasta file.
If amino acid sequences, pI and molecular weight will be calculated
}
\examples{
fasta_to_df("path/to/file.fasta", type = "AA")

}
