% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_gggenes.R
\name{flip_gggenes}
\alias{flip_gggenes}
\title{Flip and/or centre gggenes data}
\usage{
flip_gggenes(
  gggenes_df,
  gene_on,
  direction = 1,
  length = "use_max",
  centre = F
)
}
\arguments{
\item{gggenes_df}{gggenes data frame (with each gene cluster as "gene cluster")}

\item{gene_on}{Gene to centre on / use for direction}

\item{direction}{Direction for gene_on to be in [1/-1] [default = 1] (1 for right, -1 for left)}

\item{length}{Either "use_max" (use the max gene co-ordinate), or "fasta_length" -}

\item{centre}{Centre genes relative to all central ("gene_on") genes for each gene cluster in the dataframe [T/F]}
}
\value{
Data frame with each gene cluster co-ordinates set on the centre_gene and re-orientated for uniformity
}
\description{
Flip and/or centre gggenes data for clearer visualisation
}
\examples{
centre_gggenes(gggenes_df = dataframe, gene_on = "TssD", direction = -1, centre = T)
}
