% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree_function_extensions.R
\name{get_clade_nodes}
\alias{get_clade_nodes}
\title{Find internal clade nodes ancestral to named/numbered sets of tip labels (within a tree)}
\usage{
get_clade_nodes(tree, clusters_df)
}
\arguments{
\item{tree}{Tree object, e.g. read in using ape::read.tree()}

\item{clusters_df}{Dataframe with two columns: "strain", "cluster", where strain corresponds with tip-label}
}
\value{
Named vector list where the values are the internal node numbers, and the names are the cluster numbers.
}
\description{
For a tree object with an associated set of clustering (e.g. phylogroups defined by fastANI, or clustering using
fastbabs/PopPUNK), retrieve the mrca for those sets of tips alongside the fastbaps/PopPUNK cluster number,
as a named vector list.
}
\examples{
get_clade_nodes(tree_object, clusters_dataframe)
}
