% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pan_genome_extensions.R
\name{get_intersect_members}
\alias{get_intersect_members}
\title{Get members (genes) of a set intersection shown in an UpsetR plot}
\usage{
get_intersect_members(upset_df, groups, exclusive = TRUE)
}
\arguments{
\item{upset_df}{UpsetR dataframe (with the members column labelled as "Gene")}

\item{groups}{List of the groups to get the intersection for}

\item{exclusive}{TRUE/FALSE. If TRUE, only get the intersection of ONLY the groups in groups.
If FALSE, return all values shared by the groups in groups, irrespective of whether those values
are also in other groups. [Default = TRUE]}
}
\value{
List of all members in the intersection
}
\description{
Given a list of groups in a set analysis plotted in UpsetR, get all members
(i.e. Genes in this instance) in the intersection of those sets. Built to extract genes
in a panaroo dataset.
}
\examples{
get_intersect_members(dataframe_used_for_upsetR_plot, list_of_groups)
}
