% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree_function_extensions.R
\name{get_within_species_cluster_mrcas}
\alias{get_within_species_cluster_mrcas}
\title{Find internal clade nodes ancestral to named/numbered sets of tip labels within a second layer of clustering (within a tree)}
\usage{
get_within_species_cluster_mrcas(x, y, z)
}
\arguments{
\item{x}{Dataframe with two columns: "strain","phylogroup", where strain corresponds with tip-label}

\item{y}{Dataframe with two columns: "strain", "Cluster", where strain corresponds with tip-label}

\item{z}{Tree object, where the tip-labels correspond with "strain" in x and y}
}
\value{
Named vector list where the values are the internal node numbers, and the names are the cluster numbers.
}
\description{
For a tree object with an associated set of clustering (e.g. phylogroups defined by fastANI), and a finer set
of clustering (e.g. multi-level clustering of bacterial genomes using fastbaps/PopPUNK), where all of the clusters
in the latter clustering are sets entirely within the former clusters, retrieve the mrca for those sets of tips
alongside the fastbaps/PopPUNK cluster number, as a named vector list.
}
\examples{
get_within_species_cluster_mrcas(x = phylogroups_dataframe, y = clusters_dataframe, z = ggtree_object)
}
