% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenes_extensions.R
\name{gggenes_df_from_gff_dir}
\alias{gggenes_df_from_gff_dir}
\title{Create a dataframe for input to gggenes from a directory of gff3 files}
\usage{
gggenes_df_from_gff_dir(gff_dir)
}
\arguments{
\item{gff_dir}{Directory containing gff3 files}
}
\value{
Dataframe with position, orientation and description of genomic features from gff files,
in a format for plotting gene diagrams with gggenes, or further input into genoplotR dnaseqs formatting.
}
\description{
Create a dataframe for input to gggenes from a directory of gff3 files
}
\examples{
gggenes_df_from_gff_dir("path/to/gff/directory")
}
