% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenes_extensions.R
\name{gggenes_df_from_gff_file}
\alias{gggenes_df_from_gff_file}
\title{Create a dataframe for input to gggenes from a single gff3 file}
\usage{
gggenes_df_from_gff_file(gff_file)
}
\arguments{
\item{gff_file}{Gff file input}
}
\value{
Dataframe with position, orientation and description of genomic features from a single gff file,
in a format for plotting gene diagrams with gggenes, or further input into genoplotR dnaseqs formatting.
}
\description{
Create a dataframe for input to gggenes from a single gff3 file
}
\examples{
gggenes_df_from_gff_dir("path/to/gff/file")
}
