% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenes_extensions.R
\name{gggenes_df_from_gff_list}
\alias{gggenes_df_from_gff_list}
\title{Create a dataframe for input to gggenes from a list of gff3 files}
\usage{
gggenes_df_from_gff_list(gff_list)
}
\arguments{
\item{gff_list}{Directory containing gff3 files}
}
\value{
Dataframe with position, orientation and description of genomic features from gff files,
in a format for plotting gene diagrams with gggenes, or further input into genoplotR dnaseqs formatting.
}
\description{
Create a dataframe for input to gggenes from a list of gff3 files
}
\examples{
gggenes_df_from_gff_dir("path/to/gff/directory")
}
