% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree_function_extensions.R
\name{gheatmap_with_gaps}
\alias{gheatmap_with_gaps}
\title{Put a multi-column gheatmap onto a ggtree object with gaps between columns}
\usage{
gheatmap_with_gaps(
  ggtree_object,
  gheatmap_df,
  col_width = max(ggtree_object$data$x)/5,
  offset_num = 0,
  gap = col_width/10
)
}
\arguments{
\item{ggtree_object}{Tree object}

\item{gheatmap_df}{Dataframe to be put along tree as gheatmap. Rownames should correspond with the tip labels in the tree.}

\item{col_width}{Column width for each column in the heatmap [Default = 1/5 width of the tree]}

\item{offset_num}{Initial offset from the tree [Default = 0]}

\item{gap}{Gap between each column [Default = 1/10 of value of column width]}
}
\value{
Tree object with gheatmap added, with the gaps
}
\description{
Put a multi-column gheatmap onto a ggtree object with gaps between columns
}
\examples{
remove_tips_from_tree(tree_object, list_of_tips)
}
