% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{is.hex}
\alias{is.hex}
\title{Hexcode colours (from Glimma v1-hexcol.R)}
\usage{
is.hex(x, hash = TRUE)
}
\arguments{
\item{x}{the colour value(s) to check.}

\item{hash}{TRUE/FALSE - check for leading hash [Default = TRUE]}
}
\value{
Logical vector indicating if strings(s) are valid hex representations
}
\description{
Check if string(s) are valid hex colour representation
}
