% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_functions.R
\name{nested.test.check}
\alias{nested.test.check}
\title{Compare two lists of lists for nested items}
\usage{
nested.test.check(list1, list2)
}
\arguments{
\item{list1}{First list of lists to be checked}

\item{list2}{Second list of lists to be compared to}
}
\value{
list of TRUE/FALSE values stating whether lists in list1 are nested within any list in list2
}
\description{
Compare two lists comprised of lists to check whether elements of second-level lists are present entirely within other second-level lists.
}
\examples{
nested.test.function(list1,list2)
}
