% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_parsing.R
\name{parse_interproscan_gff}
\alias{parse_interproscan_gff}
\title{Parse interproscan gff3 file to a dataframe}
\usage{
parse_interproscan_gff(gff_file)
}
\arguments{
\item{gff_file}{Gff file}
}
\value{
fasta sequence as string
}
\description{
Get all domain entries for hits from interproscan into a tidy dataframe
}
\examples{
parse_interproscan_gff("input.iprscan.gff")
}
