% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_functions.R
\name{parse_snpEff_vcf}
\alias{parse_snpEff_vcf}
\title{Parse snpEff output into tidy dataframe}
\usage{
parse_snpEff_vcf(snpEff_vcf)
}
\arguments{
\item{snpEff_vcf}{First list of lists to be checked}
}
\value{
tidy dataframe
}
\description{
Parse snpEff output into tidy dataframe, using a snpEff vcf file as input
}
\examples{
parse_snpEff_vcf(snpEff_vcf)
}
