% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pan_genome_extensions.R
\name{plot_gene_accumulation_curves}
\alias{plot_gene_accumulation_curves}
\title{Plot gene accumulation curves}
\usage{
plot_gene_accumulation_curves(gene_presence_absence_file, groups)
}
\arguments{
\item{gene_presence_absence_file}{gene_presence_absence.Rtab file from Roary or Panaroo}

\item{groups}{Named vector with a list of lists of strains, with the name being each cluster / lineage / species etc.}
}
\value{
Plot of gene accumulation curves, and the dataframe used to plot these
}
\description{
Plot gene accumulation curves from a gene_presence_absence file from a pan-genome analysis
using vegan and ggplot2
}
\examples{
plot_gene_accumulation_curves(gene_presence_absence.Rtab, list_of_groups)
}
