% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoplotR_extensions.R
\name{plot_genoplot_data}
\alias{plot_genoplot_data}
\title{Plot a genoplotR synteny plot}
\usage{
plot_genoplot_data(input_genoplotR_set, output_prefix)
}
\arguments{
\item{input_genoplotR_set}{Input list of the dna_seqs and the blast comparisons}

\item{output_prefix}{Prefix for the output files (svg and png)}
}
\value{
A png and svg file of synteny plots produced using genoplotR
}
\description{
Plot a genoplotR synteny plot, using the dnaseqs from micro_gen_extra::create_dnaseqs()
and the blast comparisons from micro_gen_extra::create_blast_comparisons()
}
\examples{
plot_genoplot_data(input_genoplotR_data, "output_prefix_name")

More specifically:

#First create an empty list
genoplotr_data <- list()

#Add the dnaseqs:
genoplotr_data$dnaseqs <- micro.gen.extra::create_dnaseqs(input_gggenes_df)

#Then add the blast comparisons:
genoplotr_data$comparisons <- micro.gen.extra::create_blast_comparisons(blast_order_list, "dir_with_fasta_files")

#Then plot:
plot_genoplot_data(genoplotr_data, "output_prefix_name")

#To add colours, and change blast parameters (to be added), see the documentation for micro_gen_extra::create_dnaseqs
#and micro_gen_extra::create_blast_comparisons
}
