% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoplotR_extensions.R
\name{plot_genoplotR_key}
\alias{plot_genoplotR_key}
\title{Plot the key for a genoplotR synteny plot}
\usage{
plot_genoplotR_key(color_scheme_name = "red_blue")
}
\arguments{
\item{color_scheme_name}{Colour scheme to be used. [Default = "red_blue"]}
}
\value{
Svg plots with individual keys for forward and reverse blast hits
}
\description{
Plot the key for a genoplotR synteny plot, output as svg for both forward and reverse blast hits.
Currently only a red/blue color scheme available
}
\examples{
plot_genoplotR_key()

plot_genoplotR_key(color_scheme_name = "red_blue")
}
