% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_gggenes.R
\name{plot_t6_clusters}
\alias{plot_t6_clusters}
\title{Plot T6SS hamburger results for each strain}
\usage{
plot_t6_clusters(
  hamburger_dir,
  strains = "all",
  output_dir = ".",
  overwrite = F,
  format = "png",
  height = 8,
  width = 8
)
}
\arguments{
\item{hamburger_dir}{Hamburger output directory}

\item{strains}{List of strains to plot gene clusters from}

\item{output_dir}{Output directory to put image files}

\item{overwrite}{TRUE/FALSE for whether to overwrite output_dir (if it already exists)}

\item{format}{png/svg output, can select both - c("png","svg","both")}

\item{height}{Plot height in inches - default = 8}

\item{width}{Plot width in inches - default = 8}
}
\value{
png/svg files of gene arrow plots for extracted T6SS gene clusters using hamburger
}
\description{
Plot T6SS hamburger results for each strain
}
\examples{
gggenes_df_from_gff_dir("path/to/gff/directory")
}
