% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_parsing.R
\name{protein_from_sequence}
\alias{protein_from_sequence}
\title{Translate protein sequences from a corresponding fasta sequences, given start, end and strand}
\usage{
protein_from_sequence(fasta_sequence, start, end, strand)
}
\arguments{
\item{fasta_sequence}{String of the fasta sequence}

\item{start}{Integer of the start position of the feature to be translated in the fasta_sequence}

\item{end}{Integer of the end position of the feature to be translated in the fasta_sequence}

\item{strand}{Orientation of the feature to be translated ("+", or "-")}
}
\value{
Protein sequence
}
\description{
Translate protein sequences from a corresponding fasta sequences, given start, end and strand. Ideal for
getting all protein sequences within a gggenesdf if the fasta sequences from the gff3 file has been extracted.
}
\examples{
protein_from_sequence(List of fasta_sequence, start, end, strand)

#Particularly useful if wanting to get all the protein sequences in one/several gff3 files for comparison, e.g. running CDhit:

#E.g. Taking a bunch of gff files in a directory:




}
