% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{random_n_colours}
\alias{random_n_colours}
\title{Create a vector of "n" random colours}
\usage{
random_n_colours(num, hex = TRUE)
}
\arguments{
\item{num}{Number of colours to generate}

\item{hex}{TRUE/FALSE for hexademical colours [Default = TRUE]}
}
\value{
A vector of colours, either hexadecimal or named colours
}
\description{
Create a vector of "n" random colours using grDevices, output either hexadecimal/named colours
}
\examples{
random_n_colours(10, TRUE)
random_n_colours(num = 20, hex = FALSE)

#Names can be set to vector afterwards (e.g. to create a named set of colours to pass to ggplot2, such as scale_fill_manual())
#For example:
colours_vector <- random_n_colours(3, TRUE) # first get a vector of three hex colours
names(colours_vector) <- c("first_name","second_name","third_name") # then apply the names

Example output:
 first_name second_name  third_name
  "#CDC9A5"   "#6959CD"   "#D2691E"
}
