% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gffs.R
\name{read_gff}
\alias{read_gff}
\title{Read in annotation from gff file into a tidy dataframe}
\usage{
read_gff(gff_file)
}
\arguments{
\item{gff_file}{Gff file}
}
\value{
dataframe with annotation from gff file
}
\description{
Read in annotation from gff file into a tidy dataframe (including all fields in the decription [9th] column in gff files)
}
\examples{
read_gff("input.gff")
}
