% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree_function_extensions.R
\name{remove_tips_from_tree}
\alias{remove_tips_from_tree}
\title{Prune a list of tips from a tree}
\usage{
remove_tips_from_tree(tree, tips)
}
\arguments{
\item{tree}{Tree object}

\item{tips}{List of tips as present in tree}
}
\value{
Tree object with requested tips pruned out
}
\description{
Given a list of tips, remove these from a tree object and return the pruned tree
}
\examples{
remove_tips_from_tree(tree_object, list_of_tips)
}
