% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gffs.R
\name{reverse_gff}
\alias{reverse_gff}
\title{Reverse a gff file}
\usage{
reverse_gff(input_gff, output_gff)
}
\arguments{
\item{input_gff}{Gff file}

\item{output_gff}{Gff file}
}
\description{
Reverse a gff file and write out to file
}
\examples{
reverse_gff("path/to/input.gff", "path/to/output.gff")
}
