% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cdhit.R
\name{run_cdhit}
\alias{run_cdhit}
\title{Run CDhit on a list of protein sequences}
\usage{
run_cdhit(protein_sequences)
}
\arguments{
\item{protein_sequences}{Named vector. List comprised of protein sequences, names are the unique identifiers for the sequences.}
}
\value{
R dataframe of CDhit output
}
\description{
Run CDhit on a list of protein sequences, then read in the output into an R dataframe
}
\examples{
run_cdhit(list_of_protein_sequences)
}
