% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hmmsearch.R
\name{run_hmmsearch}
\alias{run_hmmsearch}
\title{Run Hmmsearch}
\usage{
run_hmmsearch(
  query,
  hmm_profile,
  input_format = c("fasta", "AAStringSet"),
  output_dir = "temp_hmmsearch",
  cutoff = 20,
  cpu = 1,
  clean = T
)
}
\arguments{
\item{query}{fasta file or Biostrings AAStringSet}

\item{hmm_profile}{Hmm profile file (with at least one profile hmm)}

\item{input_format}{Query input format. One of ["fasta", "AAStringSet"],}

\item{output_dir}{Output directory.}

\item{cutoff}{Hmmer score cutoff. [Default = 20]}

\item{cpu}{Number of threads to use. [Default = 1]}

\item{clean}{Remove output directory and contents. [Default = TRUE]}
}
\value{
Whole sequence and per-domain hmmsearch results as dataframes, and alignment of hits as alignment in Biostrings format
}
\description{
Run hmmsearch and store tabular data frame, domain tabular output data frame, and an alignment of the domain hits in a list
}
