% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_parsing.R
\name{snippycore_to_df}
\alias{snippycore_to_df}
\title{Read in snippy-core alignment into tidy dataframe for plotting against a tree}
\usage{
snippycore_to_df(input_file)
}
\arguments{
\item{fasta_file}{Input fasta file}
}
\value{
Tidy data frame
}
\description{
Read in snippy-core alignment (using seqinr::read.fasta) and output a tidy into a tidy dataframe
}
\examples{
snippycore_to_df("path/to/file.fasta")

}
