% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoplotR_extensions.R
\name{plot_synteny_gff_files}
\alias{plot_synteny_gff_files}
\title{Take a list of "x" gff3 files and create a synteny plot}
\usage{
plot_synteny_gff_files(
  gff_list,
  output_prefix = "synteny_plot",
  temp_dir = "temp_fasta",
  clean_up = TRUE
)
}
\arguments{
\item{gff_list}{Input list of gff files (with extensions to paths)}

\item{output_prefix}{Prefix for the output files (svg and png)}

\item{temp_dir}{Directory to put fasta files into [Default = "temp_fasta"]}

\item{clean_up}{TRUE/FALSE - clean up the temp fasta directory [Default = TRUE]}
}
\value{
A png and svg file of synteny plots produced using genoplotR
}
\description{
Take gff files, read annotation and extract fasta sequence.
Then create genoplotR data structures, then plot genoplotR synteny plot.
This uses the dnaseqs from micro_gen_extra::create_dnaseqs()
and the blast comparisons from micro_gen_extra::create_blast_comparisons()
}
\examples{
Insert examples here
}
